requirements.txt
boto3==1.13.3
PyYAML==5.1
SQLAlchemy==1.3.9
PyMySQL==0.9.3
pandas==0.25.1
xgboost==1.0.2
scikit-learn==0.21.3
Flask-SQLAlchemy==2.4.1
Flask==1.1.1
pytest==5.4.2
